abstract out the tooltip
