+++ ui/DesignSystem/Component/Sidebar.svelte
@@ -4,7 +4,8 @@
 
   import * as path from "../../src/path.ts";
 
-  import { Avatar, Icon, Title } from "../Primitive";
+  import { Tooltip } from "../Component";
+  import { Avatar, Icon } from "../Primitive";
 
   import AddOrgButton from "./Sidebar/AddOrgButton.svelte";
 
@@ -12,16 +13,6 @@
 
   export let identity = null;
   export let orgs = null;
-
-  let tooltip = { className: "hidden" };
-
-  const hideTooltip = () => {
-    tooltip.className = "hidden";
-  };
-  const showTooltip = (title, e) => {
-    const rect = e.target.closest("[data-tooltip]").getBoundingClientRect();
-    tooltip = { positionY: rect.top, title: title, className: "visible" };
-  };
 </script>
 
 <style>
@@ -108,49 +99,6 @@
     fill: var(--color-secondary);
   }
 
-  .item:hover .tooltip {
-    transition: 0.3s;
-  }
-
-  .tooltip {
-    white-space: nowrap;
-    user-select: none;
-    background-color: var(--color-foreground);
-    color: var(--color-background);
-    text-align: center;
-    border-radius: 4px;
-    padding: 4px 8px 6px 8px;
-    position: fixed;
-
-    left: 60px;
-    pointer-events: none;
-    z-index: 100;
-    margin-top: 2px;
-    margin-bottom: 2px;
-  }
-  .tooltip.visible {
-    visibility: visible;
-  }
-  .tooltip.hidden {
-    visibility: hidden;
-  }
-
-  .tooltip:before {
-    content: "";
-    display: block;
-    width: 0;
-    height: 0;
-    position: absolute;
-
-    border-top: 6px solid transparent;
-    border-bottom: 6px solid transparent;
-    border-right: 6px solid var(--color-foreground);
-    left: -6px;
-    border-top-left-radius: 30%;
-
-    top: 10px;
-  }
-
   a {
     display: flex;
     align-items: center;
@@ -160,30 +108,21 @@
   }
 </style>
 
-<div
-  style={`top: ${tooltip.positionY}px`}
-  class={`tooltip ${tooltip.className}`}>
-  <Title>{tooltip.title}</Title>
-</div>
-
 <div class="wrapper" data-cy="sidebar">
   <ul class="top">
     <li
       class="item indicator"
       data-cy="profile"
       class:active={path.active(path.profile(), $location, true)}>
-      <a
-        data-tooltip
-        on:mouseover={(e) => showTooltip(identity.metadata.handle, e)}
-        on:mouseout={hideTooltip}
-        href={path.profileProjects()}
-        use:link>
-        <Avatar
-          size="medium"
-          avatarFallback={identity.avatarFallback}
-          imageUrl={identity.metadata.avatarUrl}
-          variant="circle" />
-      </a>
+      <Tooltip value={identity.metadata.handle}>
+        <a href={path.profileProjects()} use:link>
+          <Avatar
+            size="medium"
+            avatarFallback={identity.avatarFallback}
+            imageUrl={identity.metadata.avatarUrl}
+            variant="circle" />
+        </a>
+      </Tooltip>
     </li>
 
     {#each orgs as org}
@@ -191,50 +130,40 @@
         class="item indicator"
         data-cy={`org-${org.id}`}
         class:active={path.active(path.orgs(org.id), $location, true)}>
-        <a
-          data-tooltip
-          on:mouseover={(e) => showTooltip(org.id, e)}
-          on:mouseout={hideTooltip}
-          href={path.orgProjects(org.id)}
-          use:link>
-          <Avatar
-            avatarFallback={org.avatarFallback}
-            variant="square"
-            size="medium" />
-        </a>
+        <Tooltip value={org.id}>
+          <a href={path.orgProjects(org.id)} use:link>
+            <Avatar
+              avatarFallback={org.avatarFallback}
+              variant="square"
+              size="medium" />
+          </a>
+        </Tooltip>
       </li>
     {/each}
 
     <li class="item" data-cy="add-org-button">
-      <div
-        data-tooltip
-        on:mouseover={(event) => showTooltip('Add org', event)}
-        on:mouseout={hideTooltip}>
+      <Tooltip value="Add org">
         <AddOrgButton on:click={() => dispatch('createorg')} />
-      </div>
+      </Tooltip>
     </li>
   </ul>
   <ul class="bottom">
-    <li
-      class="item indicator"
-      data-cy="wallet"
-      data-tooltip
-      on:mouseover={(e) => showTooltip('Wallet', e)}
-      on:mouseout={hideTooltip}>
-      <a href={path.profileWallet()} use:link>
-        <Icon.Fund />
-      </a>
-    </li>
-    <li
-      class="item indicator"
-      data-cy="settings"
-      class:active={path.active(path.settings(), $location)}
-      data-tooltip
-      on:mouseover={(event) => showTooltip('Settings', event)}
-      on:mouseout={hideTooltip}>
-      <a href={path.settings()} use:link>
-        <Icon.Settings />
-      </a>
-    </li>
+    <Tooltip value="Wallet">
+      <li class="item indicator" data-cy="wallet">
+        <a href={path.profileWallet()} use:link>
+          <Icon.Fund />
+        </a>
+      </li>
+    </Tooltip>
+    <Tooltip value="Settings">
+      <li
+        class="item indicator"
+        data-cy="settings"
+        class:active={path.active(path.settings(), $location)}>
+        <a href={path.settings()} use:link>
+          <Icon.Settings />
+        </a>
+      </li>
+    </Tooltip>
   </ul>
 </div>
